+++ core/utils.py
@@ -20,6 +20,10 @@ currentLanguage = ContextVar("Language", default=None)
 # Determine which ProjectID we currently run in (as the app_identity module isn't available anymore)
 _, projectID = google.auth.default()
 del _
+appVersion = os.getenv("GAE_VERSION")  # Name of this version as deployed to the appengine
+# Hash of appVersion used for cache-busting for static resources (css etc) that does not reveal the actual version name
+versionHash = urlsafe_b64encode(hashlib.sha256((appVersion+projectID).encode("UTF8")).digest()).decode("ASCII")
+versionHash = "".join([x for x in versionHash if x in string.digits+string.ascii_letters])[1:7]  # Strip +, / and =
 # Determine our basePath (as os.getCWD is broken on appengine)
 projectBasePath = globals()["__file__"].replace("/viur/core/utils.py", "")
 isLocalDevelopmentServer = os.environ['GAE_ENV'] == "localdev"